pyproject.toml
[project]
name = "onemarinex-api"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.8.2",
  "pydantic-settings==2.4.0",
  "python-multipart==0.0.9",
  "SQLAlchemy==2.0.34",
  "psycopg2-binary==2.9.9",
  "alembic==1.13.2",
  "bcrypt==4.2.0",
  "PyJWT==2.9.0",
  "boto3==1.35.21",
  "redis==5.0.8",
  "email-validator==2.2.0",
  "httpx==0.27.2"
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true
app = "app.main:app"
